+++ src/accessiweather/models/config_app.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from ..location_sorting import location_name_sort_key
 from .config_settings import AppSettings
 from .weather import Location
 
@@ -37,7 +36,7 @@ class AppConfig:
                     **({"fire_zone_id": loc.fire_zone_id} if loc.fire_zone_id else {}),
                     **({"radar_station": loc.radar_station} if loc.radar_station else {}),
                 }
-                for loc in sorted(self.locations, key=location_name_sort_key)
+                for loc in self.locations
                 if loc.name != _LEGACY_NATIONWIDE_LOCATION_NAME
             ],
             "current_location": {